refactor: merged talent and hunter routes/controllers into unified user endpoints and updated frontend fetch paths accordingly.
