+++ backend/src/app.js
@@ -1,21 +1,20 @@
 const express = require ('express');
-const cors = require ('cors');
-const talentRouter = require ('./routes/talent.routes.js');
+const cors = require('cors');
+
+const userRouter = require ('./routes/user.routes.js');
 const expRouter = require ('./routes/experiences.routes.js');
 const skillsRouter = require ('./routes/skills.routes.js');
 const uploadRouter = require('./routes/upload.route.js')
-const hunterRouter = require('./routes/hunter.route.js')
 
 const app = express();
 
 app.use(cors());
 app.use(express.json());
 
-app.use('/api/talent', talentRouter);
+app.use('/api/user', userRouter);
 app.use('/api/experiences', expRouter);
 app.use('/api/skills', skillsRouter);
 app.use('/api', uploadRouter);
-app.use('/api', hunterRouter);
 
 
 app.use((err, req, res, next) => {
