--- backend/src/routes/hunter.route.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const { getHunters } = require('../controllers/hunter.controller');
-
-const router = express.Router();
-
-router.get('/hunters', getHunters);
-
-module.exports = router;
